+++ darts/physics/super/property_container.py
@@ -151,6 +151,20 @@ class PropertyContainer(PropertyBase):
 
         return self.sat[0]
 
+    def compute_total_enthalpy(self, state, temperature):
+        _ = self.flash_ev.evaluate_PT(state[0], temperature)
+        flash_results = self.flash_ev.get_flash_results()
+        nu = np.array(flash_results.nu)
+        x = np.array(flash_results.X).reshape(self.nph, self.nc)
+
+        ph = np.array([j for j in range(self.nph) if nu[j] > 0])
+
+        enthalpy = 0.
+        for j in ph:
+            enthalpy += nu[j] * self.enthalpy_ev[self.phases_name[j]].evaluate(state[0], temperature, x[j, :])
+
+        return enthalpy
+
     def run_flash(self, pressure, temperature, zc):
         # Normalize fluid compositions
         zc_norm = zc if not self.ns else zc[:self.nc_fl] / (1. - np.sum(zc[self.nc_fl:]))
@@ -201,14 +215,19 @@ class PropertyContainer(PropertyBase):
 
         self.clean_arrays()
 
+        # Run flash
         self.ph = self.run_flash(pressure, temperature, zc)
+        self.temperature = self.flash_ev.get_flash_results().temperature if not isinstance(self.flash_ev, int) else self.temperature
+        assert self.temperature is not None, ("PropertyContainer does not specify self.temperature, should be set to "
+                                              "constant temperature in case of isothermal physics, "
+                                              "self.flash.temperature in case of thermal")
 
         for j in self.ph:
             M = np.sum(self.Mw[:self.nc_fl] * self.x[j][:self.nc_fl])
 
-            self.dens[j] = self.density_ev[self.phases_name[j]].evaluate(pressure, temperature, self.x[j, :])  # output in [kg/m3]
+            self.dens[j] = self.density_ev[self.phases_name[j]].evaluate(pressure, self.temperature, self.x[j, :])  # output in [kg/m3]
             self.dens_m[j] = self.dens[j] / M  # molar density [kg/m3]/[kg/kmol]=[kmol/m3]
-            self.mu[j] = self.viscosity_ev[self.phases_name[j]].evaluate(pressure, temperature, self.x[j, :], self.dens[j])  # output in [cp]
+            self.mu[j] = self.viscosity_ev[self.phases_name[j]].evaluate(pressure, self.temperature, self.x[j, :], self.dens[j])  # output in [cp]
         self.compute_saturation(self.ph)
 
         self.pc = self.capillary_pressure_ev.evaluate(self.sat)
@@ -219,10 +238,10 @@ class PropertyContainer(PropertyBase):
         for j in range(self.ns):
             idx = self.np_fl + j
             self.sat[idx] = zc[self.nc_fl + j]
-            self.dens[idx] = self.density_ev[self.phases_name[idx]].evaluate(pressure, temperature)
+            self.dens[idx] = self.density_ev[self.phases_name[idx]].evaluate(pressure, self.temperature)
             self.dens_m[idx] = self.dens[idx] / self.Mw[self.nc_fl + j]
 
-        self.mass_source = self.evaluate_mass_source(pressure, temperature, zc)
+        self.mass_source = self.evaluate_mass_source(pressure, self.temperature, zc)
 
         return
 
@@ -237,12 +256,12 @@ class PropertyContainer(PropertyBase):
         pressure, temperature, zc = self.get_state(state)
 
         for j in self.ph:
-            self.enthalpy[j] = self.enthalpy_ev[self.phases_name[j]].evaluate(pressure, temperature, self.x[j, :])  # kJ/kmol
-            self.cond[j] = self.conductivity_ev[self.phases_name[j]].evaluate(pressure, temperature, self.x[j, :], self.dens[j])
+            self.enthalpy[j] = self.enthalpy_ev[self.phases_name[j]].evaluate(pressure, self.temperature, self.x[j, :])  # kJ/kmol
+            self.cond[j] = self.conductivity_ev[self.phases_name[j]].evaluate(pressure, self.temperature, self.x[j, :], self.dens[j])
 
         for j in range(self.ns):
             idx = self.np_fl + j
-            self.enthalpy[idx] = self.enthalpy_ev[self.phases_name[idx]].evaluate(pressure, temperature, self.x[0, :])
+            self.enthalpy[idx] = self.enthalpy_ev[self.phases_name[idx]].evaluate(pressure, self.temperature, self.x[0, :])
             self.cond[idx] = self.conductivity_ev[self.phases_name[idx]].evaluate()
 
         # Heat source and Reaction enthalpy
@@ -251,7 +270,7 @@ class PropertyContainer(PropertyBase):
             self.energy_source += self.energy_source_ev.evaluate(state)
 
         for j, reaction in self.kinetic_rate_ev.items():
-            self.energy_source += reaction.evaluate_enthalpy(pressure, temperature, self.x, zc[self.nc_fl + j])
+            self.energy_source += reaction.evaluate_enthalpy(pressure, self.temperature, self.x, zc[self.nc_fl + j])
 
         return
 
@@ -260,10 +279,11 @@ class PropertyContainer(PropertyBase):
         pressure, temperature, zc = self.get_state(state)
 
         ph = self.run_flash(pressure, temperature, zc)
+        self.temperature = self.flash_ev.get_flash_results().temperature
 
         for j in ph:
             M = np.sum(self.Mw * self.x[j][:])  # molar weight of mixture
-            self.dens_m[j] = self.density_ev[self.phases_name[j]].evaluate(pressure, temperature, self.x[j][:]) / M
+            self.dens_m[j] = self.density_ev[self.phases_name[j]].evaluate(pressure, self.temperature, self.x[j][:]) / M
 
         self.compute_saturation(ph)