+++ engines/src/engine_nc_mp_cpu.cpp
@@ -240,13 +240,10 @@ int engine_nc_mp_cpu<NC>::init_base(conn_mesh *mesh_, std::vector<ms_well *> &we
 	new_z.resize(nc);
 	FIPS.resize(nc);
 
+	X_init = mesh->initial_state;
+
 	for (index_t i = 0; i < mesh->n_blocks; i++)
 	{
-		X_init[n_vars * i] = mesh->pressure[i];
-		for (uint8_t c = 0; c < nc - 1; c++)
-		{
-			X_init[n_vars * i + c + 1] = mesh->composition[i * (nc - 1) + c];
-		}
 		PV[i] = mesh->volume[i] * mesh->poro[i];
 		RV[i] = mesh->volume[i] * (1 - mesh->poro[i]);
 	}