+++ models/1ph_1comp_poroelastic_convergence/model.py
@@ -119,13 +119,15 @@ class Model(THMCModel):
         super().set_input_data()
 
     def set_initial_conditions(self):
+        input_distribution = {'pressure': self.reservoir.p_init}
+        input_distribution.update({comp: self.reservoir.z_init[i] for i, comp in enumerate(self.physics.components[:-1])})
         if self.reservoir.thermoporoelasticity:
-            self.physics.set_nonuniform_initial_conditions(self.reservoir.mesh,
-                                                           initial_pressure=self.reservoir.p_init,
-                                                           initial_temperature=self.reservoir.t_init,
-                                                           initial_displacement=[0.0, 0.0, 0.0])
+            input_distribution['temperature'] = self.reservoir.t_init
+            input_displacement = [0.0, 0.0, 0.0]
         else:
-            self.physics.set_nonuniform_initial_conditions(self.reservoir.mesh,
-                                                           initial_pressure=self.reservoir.p_init,
-                                                           initial_displacement=self.reservoir.u_init)
+            input_displacement = self.reservoir.u_init
+
+        self.physics.set_initial_conditions_from_array(self.reservoir.mesh,
+                                                       input_distribution=input_distribution,
+                                                       input_displacement=input_displacement)
         return 0