+++ models/2ph_constant_k/model.py
@@ -228,17 +228,25 @@ class Model(DartsModel):
         if self.reservoir_type != '1D' and self.reservoir_type != '2D':
             max_p = 1.4 * np.max(self.p_init)
             axes_max[0] = max_p
-        self.physics = Compositional(self.components, phases, self.timer, n_points=self.obl_points,
+
+        thermal = False
+        state_spec = Compositional.StateSpecification.PT if thermal else Compositional.StateSpecification.P
+        self.physics = Compositional(self.components, phases, self.timer, state_spec=state_spec, n_points=self.obl_points,
                                      min_p=40, max_p=max_p, min_z=self.zero/10, max_z=1-self.zero/10, cache=False,
                                      axes_max=axes_max)
         self.physics.add_property_region(property_container)
         
         return
 
-    def set_initial_conditions(self, initial_values: dict = None, gradient: dict = None):
+    def set_initial_conditions(self):
         if self.reservoir_type == '1D' or self.reservoir_type == '2D':
-            self.physics.set_uniform_initial_conditions(mesh=self.reservoir.mesh, uniform_pressure=self.p_init,
-                                                        uniform_composition=self.ini_comp)
+            input_distribution = {'pressure': self.p_init}
+            input_distribution.update({comp: self.ini_comp[i] for i, comp in self.physics.components[:-1]})
+            # if self.physics.thermal:
+            #     input_distribution['temperature'] = self.init_temp
+
+            return self.physics.set_initial_conditions_from_array(self.reservoir.mesh,
+                                                                  input_distribution=input_distribution)
         else:
             # get depths
             depths = np.asarray(self.reservoir.mesh.depth)
@@ -273,8 +281,8 @@ class Model(DartsModel):
                            dTdh=0.).reshape((nb, self.physics.n_vars))
 
             # assign initial condition with evaluated initialized properties
-            self.set_initial_conditions_from_depth_table(depth=init.depths,
-                                                         initial_distribution={var: X[:, i] for i, var in enumerate(self.physics.vars)})
+            self.physics.set_initial_conditions_from_depth_table(mesh=self.reservoir.mesh, input_depth=init.depths,
+                                                                 input_distribution={var: X[:, i] for i, var in enumerate(self.physics.vars)})
 
     def set_well_controls(self):
         injector = self.reservoir.get_well('I1')